refactor(#54): cleanup config and compiler
