+++ packages/pixeloven-webpack/config/src/index.test.ts
@@ -1,16 +1,16 @@
 import "jest";
 import {
-    webpackClientConfig,
-    webpackServerConfig
+    getConfig,
+    shimOptions
 } from "./index";
 
 describe("@pixeloven-webpack/config", () => {
     describe("index", () => {
-        it("should export webpackClientConfig", () => {
-            expect(typeof webpackClientConfig).toEqual("function");
+        it("should export getConfig", () => {
+            expect(typeof getConfig).toEqual("function");
         });
         it("should export webpackServerConfig", () => {
-            expect(typeof webpackServerConfig).toEqual("function");
+            expect(typeof shimOptions).toEqual("function");
         });
     });
 });
